+++ src/main/model/account/interfaces/Taxable.java
@@ -1,5 +1,7 @@
 package src.main.model.account.interfaces;
 
+import java.math.BigDecimal;
+
 public interface Taxable {
-  public void tax(double income);
+  public void tax(BigDecimal income);
 }
